download sourcode under build and install directory to <devel>/lib and <src>/share
