+++ rtctree/catkin.cmake
@@ -7,44 +7,28 @@ project(rtctree)
 find_package(catkin REQUIRED mk)
 
 # Build rtctree
-execute_process(COMMAND cmake -E chdir ${PROJECT_SOURCE_DIR} make -f Makefile.rtctree MK_DIR=${mk_PREFIX}/share/mk installed
-                RESULT_VARIABLE _make_failed)
-if (_make_failed)
-  message(FATAL_ERROR "Build of failed")
-endif(_make_failed)
-
-## System dependencies are found with CMake's conventions
-# find_package(Boost REQUIRED COMPONENTS system)
-
+# <devel>/lib/<package>/bin
+# <devel>/lib/python2.7/dist-packages
+# <src>/<package>/share
+if(NOT EXISTS ${CMAKE_CURRENT_BINARY_DIR}/installed)
+  execute_process(
+    COMMAND cmake -E chdir ${CMAKE_CURRENT_BINARY_DIR}
+    make -f ${PROJECT_SOURCE_DIR}/Makefile.rtctree installed
+    INSTALL_DIR=${CATKIN_DEVEL_PREFIX}
+    INSTALL_SCRIPTS_DIR=${CATKIN_DEVEL_PREFIX}/lib/${PROJECT_NAME}
+    INSTALL_DATA_DIR=${PROJECT_SOURCE_DIR}
+    MK_DIR=${mk_PREFIX}/share/mk
+    MD5SUM_FILE=${PROJECT_SOURCE_DIR}/rtctree-3.0.0.zip.md5sum
+    RESULT_VARIABLE _make_failed)
+  if (_make_failed)
+    message(FATAL_ERROR "Build of rtctree failed")
+  endif(_make_failed)
+endif()
 
 ## Uncomment this if the package has a setup.py. This macro ensures
 ## modules and global scripts declared therein get installed
 ## See http://ros.org/doc/api/catkin/html/user_guide/setup_dot_py.html
-catkin_python_setup()
-
-#######################################
-## Declare ROS messages and services ##
-#######################################
-
-## Generate messages in the 'msg' folder
-# add_message_files(
-#   FILES
-#   Message1.msg
-#   Message2.msg
-# )
-
-## Generate services in the 'srv' folder
-# add_service_files(
-#   FILES
-#   Service1.srv
-#   Service2.srv
-# )
-
-## Generate added messages and services with any dependencies listed here
-# generate_messages(
-#   DEPENDENCIES
-#   std_msgs  # Or other packages containing msgs
-# )
+#catkin_python_setup()
 
 ###################################
 ## catkin specific configuration ##
@@ -62,34 +46,6 @@ catkin_package(
 #  DEPENDS system_lib
 )
 
-###########
-## Build ##
-###########
-
-## Specify additional locations of header files
-## Your package locations should be listed before other locations
-# include_directories(include)
-include_directories(
-  ${catkin_INCLUDE_DIRS}
-)
-
-## Declare a cpp library
-# add_library(openrtm_tools
-#   src/${PROJECT_NAME}/openrtm_tools.cpp
-# )
-
-## Declare a cpp executable
-# add_executable(openrtm_tools_node src/openrtm_tools_node.cpp)
-
-## Add cmake target dependencies of the executable/library
-## as an example, message headers may need to be generated before nodes
-# add_dependencies(openrtm_tools_node openrtm_tools_generate_messages_cpp)
-
-## Specify libraries to link a library or executable target against
-# target_link_libraries(openrtm_tools_node
-#   ${catkin_LIBRARIES}
-# )
-
 #############
 ## Install ##
 #############
@@ -97,33 +53,9 @@ include_directories(
 # all install targets should use catkin DESTINATION variables
 # See http://ros.org/doc/api/catkin/html/adv_user_guide/variables.html
 
-## Mark executable scripts (Python etc.) for installation
-## in contrast to setup.py, you can choose the destination
-# install(PROGRAMS
-#   scripts/my_python_script
-#   DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
-# )
-
-## Mark executables and/or libraries for installation
-# install(TARGETS openrtm_tools openrtm_tools_node
-#   ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-#   LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-#   RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
-# )
-
-## Mark cpp header files for installation
-# install(DIRECTORY include/${PROJECT_NAME}/
-#   DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION}
-#   FILES_MATCHING PATTERN "*.h"
-#   PATTERN ".svn" EXCLUDE
-# )
-
-## Mark other files for installation (e.g. launch and bag files, etc.)
-# install(FILES
-#   # myfile1
-#   # myfile2
-#   DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}
-# )
+install(
+  DIRECTORY ${CATKIN_DEVEL_PREFIX}/${CATKIN_PACKAGE_PYTHON_DESTINATION}/
+  DESTINATION ${CATKIN_PACKAGE_PYTHON_DESTINATION})
 
 #############
 ## Testing ##
