+++ rtctree/manifest.xml
@@ -18,7 +18,7 @@
   <rosdep name="python-omniorb" />
 
   <export>
-    <python path="${prefix}/lib/python2.7/site-packages" />
+    <python path="${prefix}/lib/python2.7/dist-packages" />
   </export>
 
 </package>
