+++ rtshell/Makefile.rtshell
@@ -1,4 +1,6 @@
 INSTALL_DIR = ${CURDIR}
+INSTALL_SCRIPTS_DIR = ${CURDIR}
+INSTALL_DATA_DIR    = ${CURDIR}
 SOURCE_DIR  = build/rtshell-3.0.0
 TARBALL     = build/rtshell-3.0.0.zip
 TARBALL_URL = https://github.com/gbiggs/rtshell/archive/3.0.0.zip
@@ -10,7 +12,7 @@ UNPACK_CMD  = unzip
 include $(MK_DIR)/download_unpack_build.mk
 
 installed: $(SOURCE_DIR)/unpacked
-	(cd $(SOURCE_DIR) && yes | python setup.py build && yes | python setup.py install --prefix $(INSTALL_DIR) --record installed_files.txt)
+	(cd $(SOURCE_DIR) && yes | python setup.py build && yes | python setup.py install --prefix $(INSTALL_DIR) --install-scripts $(INSTALL_SCRIPTS_DIR) --install-data $(INSTALL_DATA_DIR) --install-layout=deb --record installed_files.txt)
 	touch installed
 
 clean:
