+++ rtshell/manifest.xml
@@ -22,7 +22,7 @@
   <rosdep name="mk" />
 
   <export>
-    <python path="${prefix}/lib/python2.7/site-packages" />
+    <python path="${prefix}/lib/python2.7/dist-packages" />
   </export>
 
 </package>
