--- rtshell/test/run_rtshell.test
@@ -1,3 +1,0 @@
-<launch>
-   <test test-name="run_rtshell" pkg="rtshell" type="run_rtshell_test.py" />
-</launch>
